remove fasthttp2 due to instability and number of bugs. new feature of parallel reqs with http/2 per connection as http/2 can send multiple reqs on same conn using streams

+++ pkgs/payloader/output/cli/cli.go
@@ -62,7 +62,7 @@ func displayRespSize(resp payloader.ByteSize, t table.Writer) {
 	t.AppendSeparator()
 }
 
-func displayErrors(errors map[string]uint, t table.Writer) {
+func displayErrors(errors map[string]uint64, t table.Writer) {
 	rows := make([]table.Row, 0)
 	for err, count := range errors {
 		rows = append(rows, table.Row{"Error; " + err, count})
